Fix padding issue on some targets
GCC reports `offset of packed bit-field ‘green’ has changed in GCC 4.4`,
because there could be padding between uint8_t on some targets prior to
gcc 4.4 [1]. Switch to uint16_t so there will be no unintentional
padding.

[1] https://www.gnu.org/software/gcc/gcc-4.4/changes.html

Fixes #10

+++ src/utils.c
@@ -269,9 +269,9 @@ int alock_grayscale_image(XImage *image,
 
     union {
         struct __attribute__ ((packed)) {
-            uint8_t red   : 5;
-            uint8_t green : 6;
-            uint8_t blue  : 5;
+            uint16_t red   : 5;
+            uint16_t green : 6;
+            uint16_t blue  : 5;
         } v16;
         struct __attribute__ ((packed)) {
             uint8_t red;
